AdamW

+++ src/only_std/extract_dataset.rs
@@ -2,8 +2,8 @@ use std::fs::File;
 use std::io::Read;
 
 // load images from MNIST dataset
-pub fn mnist_input(path: String) -> Result<Vec<Vec<f64>>, std::io::Error>{
-    let mut inputs: Vec<Vec<f64>> = Vec::new();
+pub fn mnist_input(path: String) -> Result<Vec<Vec<f32>>, std::io::Error>{
+    let mut inputs: Vec<Vec<f32>> = Vec::new();
 
     let mut file: File = File::open(path)?;
     let mut buffer: Vec<u8> = Vec::new();
@@ -12,9 +12,9 @@ pub fn mnist_input(path: String) -> Result<Vec<Vec<f64>>, std::io::Error>{
 
     let mut i: usize = 16;
     while i < buffer.len(){
-        let mut curr_input: Vec<f64> = Vec::new();
+        let mut curr_input: Vec<f32> = Vec::new();
         for _ in 0..784{
-            curr_input.push(buffer[i] as f64 / 255.0);
+            curr_input.push(buffer[i] as f32 / 255.0);
             i += 1;
         }
         inputs.push(curr_input);
@@ -24,8 +24,8 @@ pub fn mnist_input(path: String) -> Result<Vec<Vec<f64>>, std::io::Error>{
 }
 
 // load labels for images from MNIST dataset
-pub fn mnist_output(path: String) -> Result<Vec<Vec<f64>>, std::io::Error>{
-    let mut outputs: Vec<Vec<f64>> = Vec::new();
+pub fn mnist_output(path: String) -> Result<Vec<Vec<f32>>, std::io::Error>{
+    let mut outputs: Vec<Vec<f32>> = Vec::new();
 
     let mut file: File = File::open(path)?;
     let mut buffer: Vec<u8> = Vec::new();
